decomp: sweep wins across CharBones/LayerDir/BandCharDesc/TourDescPanel (+~2.7pp)
- CharBones::RotateTo 93.9 → 95.6% — FMA fission: hoist sqw*f into temp
  inside ByteQuat/ShortQuat branches to defeat fmadds contraction
- CharBonesSamples::EvaluateChannel 90.7 → 90.8% — int q1/q2 decl swap
- LayerDir::RefreshLayer 84.2 → 84.7% — flip operand order on size
  comparison (arr->Size() > N → N < arr->Size())
- BandCharDesc::ComputeDeformWeights 76.2 → 76.5% — reorder loop locals,
  rewrite blend as plain L→R, swap muscle/weight clamp order
- TourDescProvider::UpdateExtendedCustom 84.23 → 84.46% — add `TheTour &&`
  guard before bronze.pic strcmp (null_guard_insert)
- Mtx.h ByteQuat::Set + CharBones.h MakeShortAng/ToShort — add (float)
  cast between floor() and (T) cast to force frsp before fctiwz
- Document the (T)(float)floor(...) frsp-forcing pattern in
  docs/decomp/patterns/fixable-casting.md

+++ src/system/math/Mtx.h
@@ -262,10 +262,10 @@ class ByteQuat {
 public:
     char x, y, z, w;
     void Set(const Hmx::Quat &q) {
-        x = (char)floor(Clamp(-127.0f, 127.0f, 0.5f + q.x * 127.0f));
-        y = (char)floor(Clamp(-127.0f, 127.0f, 0.5f + q.y * 127.0f));
-        z = (char)floor(Clamp(-127.0f, 127.0f, 0.5f + q.z * 127.0f));
-        w = (char)floor(Clamp(-127.0f, 127.0f, 0.5f + q.w * 127.0f));
+        x = (char)(float)floor(Clamp(-127.0f, 127.0f, 0.5f + q.x * 127.0f));
+        y = (char)(float)floor(Clamp(-127.0f, 127.0f, 0.5f + q.y * 127.0f));
+        z = (char)(float)floor(Clamp(-127.0f, 127.0f, 0.5f + q.z * 127.0f));
+        w = (char)(float)floor(Clamp(-127.0f, 127.0f, 0.5f + q.w * 127.0f));
     }
     void ToQuat(Hmx::Quat& q) const {
         q.Set(
